Feature: Handle Confirm Order

+++ admin-dashboard/src/api/axios.js
@@ -24,8 +24,8 @@ axiosInstance.interceptors.request.use((res) => {
     return res;
 }, (error) => {
     console.log(error.response);
-    const { status } = error.response;
-    if(status === 500){
+    const status = error.response ? error.response.status : 500;
+    if(status && status === 500){
         localStorage.clear();
         store.dispatch({
             type: authConstants.LOGOUT_SUCCESS
